chore: refactor post feature implementation
