+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gavdi/cap-mcp",
-  "version": "0.10.2-alpha.17",
+  "version": "0.10.1",
   "description": "MCP Pluging for CAP",
   "keywords": [
     "MCP",
